Implement suggested changes in PR
